Add filter support to GraphQL Controller
Pulls out some non-controller stuff into a series of filters. This will also allow naptime clients to specify their own filters, which can be used for more implementation-specific things (i.e. reporting to Datadog), implementing better security restrictions, etc.
