+++ examples/src/main/scala/ResourceModule.scala
@@ -9,6 +9,8 @@ import org.coursera.naptime.ari.engine.LoggingEngineMetricsCollector
 import org.coursera.naptime.ari.fetcher.LocalFetcher
 import org.coursera.naptime.ari.graphql.DefaultGraphqlSchemaProvider
 import org.coursera.naptime.ari.graphql.GraphqlSchemaProvider
+import org.coursera.naptime.ari.graphql.controllers.filters.DefaultFilters
+import org.coursera.naptime.ari.graphql.controllers.filters.FilterList
 import resources.UserStore
 import resources.UserStoreImpl
 import resources.UsersResource
@@ -29,5 +31,6 @@ class ResourceModule extends NaptimeModule {
     bind[EngineMetricsCollector].to[LoggingEngineMetricsCollector]
     bind[SchemaProvider].to[LocalSchemaProvider]
     bind[GraphqlSchemaProvider].to[DefaultGraphqlSchemaProvider]
+    bind[FilterList].to[DefaultFilters]
   }
 }
