+++ naptime-graphql/src/main/scala/org/coursera/naptime/ari/graphql/controllers/filters/DefaultFilters.scala
@@ -1,0 +1,12 @@
+package org.coursera.naptime.ari.graphql.controllers.filters
+
+import javax.inject.Inject
+import javax.inject.Singleton
+
+@Singleton
+class DefaultFilters @Inject() (
+    queryComplexityFilter: QueryComplexityFilter,
+    engineMetricsFilter: EngineMetricsFilter)
+  extends FilterList(List(
+    queryComplexityFilter,
+    engineMetricsFilter))
