+++ naptime-graphql/src/main/scala/org/coursera/naptime/ari/graphql/controllers/filters/Filter.scala
@@ -1,0 +1,12 @@
+package org.coursera.naptime.ari.graphql.controllers.filters
+
+import scala.concurrent.Future
+
+trait Filter {
+
+  type FilterFn = IncomingQuery => Future[OutgoingQuery]
+
+  def apply(nextFilter: FilterFn): FilterFn
+
+}
+
